RDKDEV-257: IPower: update the StateChange callback function to have more context
This commit is for changing the callback function to notify the power
state change to have more context.
A new struct 'PowerState' is introduced which will have the details of
previous power state, new power state and the phase(Before/After).
PowerState instance is passed as an argument for StateChange().
This change helps to send before and after power state change notification.
It also helps the observers to take necessary actions depending on the previous/current state.