RDKDEV-257: IPower: Update StateChange callback function to have more context
This commit adds more context to the callback function that notify the power state change events.
In addition to the new power state, the current power state and
the phase of transition (Before/After) is also included in parameters.
This change helps to send "Before" and "After" power state change notification.