[POC] How to define a firebolt API on through the Thunder Contract definitions.
As this is a show-and-tell to analyse the different angels of developping interfaces in a
Thudner cmpliant way, we will be comparing notes to aalyse do's and don'ts on interface design
through discussion. Using he fullest of the Code generators available today to generate a
COMRPC proxy/stubs and JSONRPC boiler plated code.

GOAL:
Avoid implementation details and use as much as possible the generated code so implementation
is as simple as possible.

Proposal by Pierre Wielders.